requirements.txt
sentence-transformers
langchain
flask
pypdf
python-dotenv
pinecone-client[grpc]==3.2.2
langchain-pinecone==0.1.0
langchain_community
langchain_openai
langchain_experimental
openai
huggingface_hub
gunicorn
flask_cors
bs4
google-search-results
-e .
